+++ core/grid.h
@@ -8,11 +8,11 @@
 
 typedef struct grid {
     TrnTetrominoType** tetrominoTypes;
-    unsigned int numberOfRows;
-    unsigned int numberOfColumns;
+    int numberOfRows;
+    int numberOfColumns;
 } TrnGrid;
 
-TrnGrid* trn_grid_new(unsigned int numberOfRows, unsigned int numberOfColumns);
+TrnGrid* trn_grid_new(int numberOfRows, int numberOfColumns);
 void trn_grid_destroy(TrnGrid* grid);
 
 void trn_grid_clear(TrnGrid* grid);
@@ -35,7 +35,7 @@ void trn_grid_copy_row_bellow(TrnGrid* grid, int rowIndex);
 
 void trn_grid_pop_row_and_make_above_fall(TrnGrid* grid, int rowIndexToPop);
 
-size_t trn_grid_pop_first_complete_rows_block();
+int trn_grid_pop_first_complete_rows_block();
 
 int tnr_grid_find_last_complete_row_index(TrnGrid* grid);
